Add Phoenix socket protocol version and versioned Message decoding/encoding support.
This patch allows the PhoenixSocket to work with older Phoenix Framework
installations (e.g. 1.3.4) that are likely to use version 1 of the
socket protocol.

In addition the implemntation is now more generally aware of the Phoenix
socket protocol version it was configured with, thus allowing future
protocol changes without breaking older and already working
implementation support.